requirements.txt
# MedGemma Medical AI — Python dependencies
# Pin ALL versions for reproducibility in a medical AI project.
# Update versions deliberately (test before updating in production).

# ── Core ML / HAI-DEF models ─────────────────────────────────────────────────
torch>=2.2.0,<3.0
# transformers: MedASR requires the dev build at a specific commit.
# The stable 4.x and 5.2.x builds have a LasrFeatureExtractor arity bug.
# Install with:
#   pip install git+https://github.com/huggingface/transformers.git@65dc261512cbdb1ee72b88ae5b222f2605aad8e5
# Listed here as >=5.0.0.dev0 so pip does not downgrade to 4.x.
transformers>=5.0.0.dev0
accelerate>=0.27.0,<2.0
peft>=0.9.0,<2.0           # LoRA fine-tuning for MedGemma
huggingface_hub>=1.0.0     # 0.x has breaking API changes vs 1.x

# ── Medical imaging ──────────────────────────────────────────────────────────
pydicom>=2.4.0,<3.0        # DICOM loading and tag manipulation
SimpleITK>=2.3.0,<3.0      # CT volume loading, resampling
openslide-python>=1.3.0    # Whole-slide image (WSI) pathology
Pillow>=10.0.0             # Image loading and manipulation
opencv-python>=4.9.0       # Computer vision (Otsu thresholding)
albumentations>=1.4.0      # Medical image augmentation

# ── Audio processing ─────────────────────────────────────────────────────────
librosa>=0.10.0,<1.0       # Audio loading, resampling
soundfile>=0.12.0          # Audio file I/O (WAV)

# ── Google Cloud (MedASR) ────────────────────────────────────────────────────
google-cloud-speech>=2.26.0  # MedASR via Google Cloud Speech-to-Text

# ── Clinical data standards ──────────────────────────────────────────────────
fhir.resources>=7.0.0     # HL7/FHIR resource parsing

# ── Clinical NLP ─────────────────────────────────────────────────────────────
spacy>=3.7.0               # NLP pipeline
# scispacy>=0.5.4          # Scientific/medical NLP (install separately — large)
# en_core_sci_lg: install via:
#   pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_lg-0.5.4.tar.gz

# ── PHI de-identification ────────────────────────────────────────────────────
presidio-analyzer>=2.2.0   # Microsoft Presidio PHI detection
presidio-anonymizer>=2.2.0

# ── Drug databases / pharmacology ────────────────────────────────────────────
requests>=2.31.0           # RxNorm API, OpenFDA API calls
# rdkit                    # SMILES validation for TxGemma (install separately)

# ── API / web ────────────────────────────────────────────────────────────────
fastapi>=0.110.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
python-multipart>=0.0.9    # File uploads
pydantic>=2.0.0,<3.0

# ── Config ───────────────────────────────────────────────────────────────────
pyyaml>=6.0.1

# ── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0              # FastAPI test client
